Use actually resolvable dependencies for native libraries
rpm doesn't really care about paths or file names. It deals with capabilities.
A libraries capabilities are the SONAMEs it provides. Those can be found using
rpm's find-provides script. The script's output is directly usable for
"Requires" fields in a spec file.

Unfortunately this works only if the dependency actually specifies the API
version of the required native library as often unversioned symlinks (like
/usr/lib/libfoo.so) are only provided by -devel packages. On the other hand
expecting to be compatible with some random library version is a rather risky
course anyway, so it's a good idea to push our module authors to explicitly
state their requirements.